requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
datadog-api-client==2.19.0
google-generativeai>=0.8.0
numpy==1.26.2
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
